% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation_functions.r
\name{ari_function}
\alias{ari_function}
\title{Add ARI Function}
\usage{
ari_function(clusters, ground_truth)
}
\arguments{
\item{clusters}{Dataframe column containing clustering assignments}

\item{ground_truth}{Dataframe column containing ground_truth assignments}
}
\description{
Function to run ARI
}
\keyword{ARI}
