% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{baron_mouse_1}
\alias{baron_mouse_1}
\title{Baron Mouse 1 Dataset}
\format{
Expression matrix with cells as columns and rows as genes
}
\usage{
baron_mouse_1
}
\description{
Dataset containing
}
\keyword{baron_mouse_1}
