% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{baron_mouse_1_ground_truth}
\alias{baron_mouse_1_ground_truth}
\title{Baron Mouse 1 Ground Truth}
\format{
Dataframe containing in column "clust_assign" the ground truth clusters for the Baron Mouse 1 Dataset
}
\usage{
baron_mouse_1_ground_truth
}
\description{
Dataset containing
}
\keyword{baron_mouse_1_ground_truth}
