% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chai.R
\name{chai}
\alias{chai}
\title{Conditional Hypothesis testing using Auxiliary Information (chai) main function}
\usage{
chai(z, X, K_vec = 2:6, R = 100)
}
\arguments{
\item{z}{A numeric vector that saving the z-statistics.}

\item{X}{A numeric matrix/vector (covariates/side-information) that corresponding to z.}

\item{K_vec}{An integer value or a range specifying the numbers of mixture components. Suggest use 2 or above. If there is a vector/range, the model will automatically select the "best" based on BIC. The default is K_vec = 2:6.}

\item{R}{An integer value indicating the total number of samples the model generate to estimate the \eqn{\pi_0(x)}. The default is R = 100. More details please see the original paper.}
}
\value{
A list with the following components:
\item{z}{The input z-statistics.}
\item{X}{The input X (side information).}
\item{K}{The optimal number of mixture components.}
\item{R}{The input R value.}
\item{clFDR}{The conditional local FDR  of every hypotheses.}
\item{pi0}{The estimated \eqn{\pi_0(x)} of every hypotheses.}
\item{post_w}{The posterior weight \eqn{w_{ik}} of every hypotheses (\eqn{i}) belong to which component (\eqn{k}).}
\item{ord}{The order of Hypothesis indices by increasing conditional local FDR (from smallest to largest).}
\item{clFDR_sorted}{The sorted conditional local FDR (from smallest to largest).
}
\description{
Conditional Hypothesis testing using Auxiliary Information (chai) main function
}
