% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{CHAI_SNF}
\alias{CHAI_SNF}
\title{Creates a similarity matrix and does clustering for snf_matrix
To be used internally for the create_matrix_list function}
\usage{
CHAI_SNF(sce, best_k, eval = TRUE)
}
\arguments{
\item{sce}{SingleCellExperiment object containing all of the clustering assignments in colData}

\item{best_k}{Either is the max_k to evaluate k until, or the true_k in the dataset, depending on if EVAL is true}

\item{eval}{if eval is true, evaluate the best_k, else use best_k to be true_k}
}
\value{
sce
}
\description{
Creates a similarity matrix and does clustering for snf_matrix
To be used internally for the create_matrix_list function
}
