% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chai_st.r
\name{CHAI_ST}
\alias{CHAI_ST}
\title{Runs CHAI with ST integration with SNF - Two Level}
\usage{
CHAI_ST(st_matrix, sce)
}
\arguments{
\item{st_matrix}{binary ST matrix}

\item{sce}{SingleCellExperiment object with all clustering assignments.}
}
\value{
adj
}
\description{
Runs CHAI with ST integration with SNF - Two Level
}
