% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clFDRselect.R
\name{clfdrselect}
\alias{clfdrselect}
\title{Conditional local FDR-based selection.}
\usage{
clfdrselect(c, q = 0.05, max_clFDR = 1)
}
\arguments{
\item{c}{A chai output or a numeric vector of conditional local FDR.}

\item{q}{The target FDR level. The default is q = 0.05.}

\item{max_clFDR}{The maximum allowable threshold when selecting rejections based on the target FDR level.
The default is max_clFDR = 1, meaning that no additional filtering is applied to the selected clFDR values.
In extreme cases, when there are many very small clFDR values, some relatively large clFDR values may also be rejected due to compensation in the average.
In such situations, it may be worth considering lowering max_clFDR.}
}
\value{
An index vector indicating the indices of the rejected hypotheses, ordered by clFDR from smallest to largest.
}
\description{
This applies to chai output or estimated clFDR values.

The procedure is as follows: we first sort the clFDR values in ascending order, and then define the number of rejection \eqn{l} as:

\eqn{L = \max \left\{ l \in \{1, \dots, m\} \,:\, \frac{1}{l} \sum_{i=1}^{l} \text{clfdr}_{(i)} \leq q\right\}}

where \eqn{q} is the desired FDR level.
}
\seealso{
\code{\link[chai]{chai}} for the Conditional Hypothesis testing using Auxiliary Information (chai) function,
\code{\link[performance]{performance}} for checking the model performance with known ground truth.
\code{\link[direction]{direction}} for the direction of the rejected hypotheses (binary outcomes only).
}
