% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avgsim_clustering.r, R/cspa_clustering.r
\name{create_average_matrix}
\alias{create_average_matrix}
\title{Average Similarity Matrix}
\usage{
create_average_matrix(similarity_matrix_list)

create_average_matrix(similarity_matrix_list)
}
\arguments{
\item{similarity_matrix_list}{list of binary similarity matrices}
}
\description{
Function to create average similarity matrix for CHAI_AvgSim

Function to create average similarity matrix for CHAI_AvgSim
}
\examples{
create_average_matrix(similarity_matrix_list)
create_average_matrix(similarity_matrix_list)
}
\keyword{Average}
\keyword{Matrix}
\keyword{Similarity}
