% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_assignment_files.r
\name{create_matrix}
\alias{create_matrix}
\title{Creates a Binary Similarity Matrix for Clustering Assignments. Internal Function.}
\usage{
create_matrix(assignment_col)
}
\arguments{
\item{assignment_col}{column containing clustering assignments}
}
\description{
To be used internally for the create_matrix_list function
}
\examples{
\dontrun{
create_matrix_list(getwd())
}
}
\keyword{Binary}
\keyword{Creation}
\keyword{Matrix}
\keyword{Similarity}
