% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_assignment_files.r
\name{create_matrix_list}
\alias{create_matrix_list}
\title{Create List of Similarity Matrices based on Clustering Assignments}
\usage{
create_matrix_list(home_dir = getwd())
}
\arguments{
\item{home_dir}{Directory where subdirectories are stored, such as ground truth, alg_clust_assign, data.}
}
\description{
Create Similarity Matrices for each clustering assignment and add it to a list. Parses the "alg_clust_assign" directory. List can be used for both CSPA and SNF
}
\examples{
create_matrix_list(getwd())
}
\keyword{Creation}
\keyword{List}
\keyword{Matrix}
\keyword{Similarity}
