% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snf_clustering.r
\name{create_snf_matrix}
\alias{create_snf_matrix}
\title{SNF Matrix}
\usage{
create_snf_matrix(similarity_matrix_list)
}
\arguments{
\item{similarity_matrix_list}{list of binary similarity matrices}
}
\description{
Function to create SNF matrix
}
\examples{
\dontrun{
snf_matrix <- create_snf_matrix(similarity_matrix_list)
}
}
\keyword{Matrix}
\keyword{SNF}
