% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/direction.R
\name{direction}
\alias{direction}
\title{Inspect the direction of rejected hypotheses (binary outcomes only)}
\usage{
direction(outcome, z_value, selected_idx)
}
\arguments{
\item{outcome}{An binary factor outcome of length n.
Please ensure that its ordering matches the sample order used when the z-statistics were originally computed.}

\item{z_value}{A numberic vector of length m containing the z-statistics.}

\item{selected_idx}{An index vector containing the indices of the hypotheses rejected by the model.
Please ensure that the indices are aligned with the ordering of the z-statistics.}
}
\value{
A data frame with the following columns:
\item{feature}{The feature names of the selected rejected hypotheses, taken from the names of the \code{z_value} input.
If names are available, they are retained; otherwise, the indices are used.}
\item{z_value}{The corresponding z-statistic values.}
\item{direction}{The direction of each selected rejected hypothesis with respect to the binary outcome.}
}
\description{
This function is designed to quickly determine the direction of the hypotheses rejected by the model when the outcome is binary.
It is recommended for use together with \code{\link[chai]{chai}} and \code{\link[clfdrselect]{clfdrselect}}.
Based on the sign of the z-statistics, the function identifies each rejected feature as either higher or lower in level 1 of the outcome.
}
\seealso{
\code{\link[chai]{chai}} for the Conditional Hypothesis testing using Auxiliary Information (chai) function,
\code{\link[clfdrselect]{clfdrselect}} for the rejected hypotheses identified,
\code{\link[performance]{performance}} for checking the model performance with known ground truth.
}
