% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation_functions.r
\name{evaluation_table}
\alias{evaluation_table}
\title{Create Eval Table}
\usage{
evaluation_table(sce, ground_truth)
}
\arguments{
\item{sce}{SingleCellExperiment object where clustering assignments are stored in colData}

\item{ground_truth}{Dataframe column containing ground_truth assignments}
}
\description{
Function to run ARI
}
\examples{
\dontrun{
evaluate_table(sce, ground_truth$clust_assign)
}
}
\keyword{NMI}
