% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_clustering_functions.r
\name{get_clust_assignments}
\alias{get_clust_assignments}
\title{Run All Clustering Assignment Algorithms currently in the pipeline}
\usage{
get_clust_assignments(sce, n_cores = 1, svm_max = 500, max_k = 15)
}
\arguments{
\item{sce}{SingleCellExperiment object of the dataset}

\item{n_cores}{Computer cores for parallel library allocated for running functions.}

\item{svm_max}{define the maximum number of cells below which SVM is not run.}

\item{max_k}{maxk range for sc3}
}
\description{
Function to run all clustering algorithms (Seurat, scSHC, Spectrum, RaceID, SC3).
}
\examples{
get_clust_assignments(sce)
}
\keyword{Algorithm}
\keyword{All}
\keyword{Function}
\keyword{Wrapper}
