% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation_functions.r
\name{nmi_function}
\alias{nmi_function}
\title{NMI Function}
\usage{
nmi_function(clusters, ground_truth)
}
\arguments{
\item{clusters}{Dataframe column containing clustering assignments}

\item{ground_truth}{Dataframe column containing ground_truth assignments}
}
\description{
Function to run NMI
}
\examples{
nmi_function(snf_clusters@.Data, ground_truth$clust_assign)
}
\keyword{NMI}
