% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance.R
\name{performance}
\alias{performance}
\title{Check model performance with known ground truth}
\usage{
performance(gt, sel, z = NULL, bh_level = 0.05, one.sided = TRUE)
}
\arguments{
\item{gt}{An index vector that saving ground truth indices.}

\item{sel}{A index vector that saving the indices of the selected hypotheses.}

\item{z}{A numeric vector that saving the z-statistics.
This only needs to be provided when Benjamini-Hochberg (BH) performance is also being evaluated.
The default is set to NULL.}

\item{bh_level}{The target FDR level for Benjamini-Hochberg (BH) performance. The default is bh_level = 0.05.}

\item{one.sided}{A Boolean operators indicating whether the p-value used to evaluate BH performance
are computed from the z-statistics using a one-sided or two-sided test.}
}
\value{
A list with the following components:
\item{FDP}{The false discovery proportion calculated based on the ground truth and selections.}
\item{Power}{The power calculated based on the ground truth and selections.}
\item{True_Signal}{The number or true signals from the ground truth.}
\item{Selected}{The number of selections.}
\item{True_Positive}{The number of selections that are true signals.}
\item{False_Positive}{The number of selections that are NOT true signals, only in selected.}
\item{False_Negative}{The number of true signals that were NOT selected, only in ground truth.}
\item{BH_FDP}{The false discovery proportion calculated based on the ground truth and BH selections.}
\item{BH_Power}{The power calculated based on the ground truth and BH selections.}
\item{BH_True_Positive}{The number of BH selections that are true signals.}
\item{BH_False_Positive}{The number of BH selections that are NOT true signals.}
\item{BH_False_Negative}{The number of true signals that were NOT selected by BH.}
\item{BH_Selected}{The number of BH selections.}
}
\description{
Check model performance with known ground truth
}
\examples{
\dontrun{
# Generate a simulation
set.seed(123)
n = 1000; n0 <- 950; n1 <- 50
z0 <- rnorm(n0, mean = 0, sd = 1)
x0 <- rnorm(n0, mean = 3, sd = 1)
z1 <- rnorm(n1, mean = 3, sd = 1)
x1 <- rnorm(n1, mean = 6, sd = 1)
z <- c(z0, z1)
X <- c(x0, x1)
gt <- seq((n0+1), n)

# Fit the model
res <- chai(z, X, K_vec = 2:6, R = 100)

# Check the rejection indices
clfdrselect(res$clFDR, q = 0.05)

# Check performance with ground truth
performance(gt, clfdrselect(res$clFDR, q = 0.05))
}
}
\seealso{
\code{\link[chai]{chai}} for the Conditional Hypothesis testing using Auxiliary Information (chai) function,
\code{\link[clfdrselect]{clfdrselect}} for the rejected hypotheses identified,
\code{\link[direction]{direction}} for the direction of the rejected hypotheses (binary outcomes only).
}
