% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_clustering_functions.r
\name{raceid_assign_func}
\alias{raceid_assign_func}
\title{Run RaceID}
\usage{
raceid_assign_func(sce, out_dir = getwd())
}
\arguments{
\item{sce}{SingleCellExperiment object of the dataset}

\item{out_dir}{Directory where output should be written to. Defaults to the current working directory. Will automatically create a directory called 'alg_clust_assign' to house all algorithm assignments}
}
\description{
Function to run RaceID clustering algorithm.
}
\examples{
raceid_assign_func(sce, out_dir = getwd())
}
\keyword{RaceID}
