% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral_clustering.r
\name{spectral_clustering}
\alias{spectral_clustering}
\title{Run Spectral Clustering}
\usage{
spectral_clustering(matrix, best_k)
}
\arguments{
\item{matrix}{Either CSPA or SNF matrix}

\item{best_k}{Number of clusters that should be run for spectral clustering}
}
\description{
On either CSPA or SNF matrix, run spectral clustering with user defined k.
}
\examples{
best_k <- calc_silhouette_score(matrix, 15)
}
\keyword{Clustering}
\keyword{Run}
\keyword{Spectral}
