% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chai_st.r
\name{CHAI_ST}
\alias{CHAI_ST}
\title{Runs CHAI with ST integration with SNF - Two Level}
\usage{
CHAI_ST(st_matrix, similarity_matrix_list)
}
\arguments{
\item{st_matrix}{binary ST matrix}

\item{similarity_matrix_list}{similarity matrix list created from clustering assignment list}
}
\value{
adj
}
\description{
Runs CHAI with ST integration with SNF - Two Level
}
