% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_clustering_functions.r
\name{choir_assign_func}
\alias{choir_assign_func}
\title{Run Choir}
\usage{
choir_assign_func(sce, n_cores = 1)
}
\arguments{
\item{sce}{SingleCellExperiment object of the dataset}
}
\description{
Function to run choir clustering algorithm
}
\examples{
choir_assign_func(sce, out_dir = getwd())
}
\keyword{choir}
