% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cspa_clustering.r
\name{create_average_matrix}
\alias{create_average_matrix}
\title{Average Similarity Matrix}
\usage{
create_average_matrix(similarity_matrix_list)
}
\arguments{
\item{similarity_matrix_list}{list of binary similarity matrices}
}
\description{
Function to create average similarity matrix for CSPA
}
\examples{
create_average_matrix(similarity_matrix_list)
}
\keyword{Average}
\keyword{CSPA}
\keyword{Matrix}
