% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation_functions.r
\name{evaluation_table}
\alias{evaluation_table}
\title{Create Eval Table}
\usage{
evaluation_table(sce, ground_truth)
}
\arguments{
\item{ground_truth}{Dataframe column containing ground_truth assignments}

\item{clusters}{Dataframe column containing clustering assignments}
}
\description{
Function to run ARI
}
\examples{
evaluate_assignments(snf_clusters@.Data, ground_truth$clust_assign)
}
\keyword{NMI}
