% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_clustering_functions.r
\name{raceid_assign_func}
\alias{raceid_assign_func}
\title{Run RaceID}
\usage{
raceid_assign_func(sce)
}
\arguments{
\item{sce}{SingleCellExperiment object of the dataset}
}
\description{
Function to run RaceID clustering algorithm.
}
\examples{
raceid_assign_func(sce, out_dir = getwd())
}
\keyword{RaceID}
