% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_clustering_functions.r
\name{sc3_assign_func}
\alias{sc3_assign_func}
\title{Run SC3}
\usage{
sc3_assign_func(sce, n_cores = 1, svm_max = 1000, max_k = 15)
}
\arguments{
\item{sce}{SingleCellExperiment object of the dataset}

\item{n_cores}{Computer cores for parallel library allocated for running functions.}

\item{svm_max}{Max number of cells to use for SVM. Default is 1000}

\item{max_k}{max_k for sc3 to run to determine optimal clustering.}
}
\description{
Function to run SC3 clustering algorithm.
}
\examples{
\dontrun{
sce3_assign_func(sce)
}
}
\keyword{SC3}
