% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_clustering_functions.r
\name{sc3_assign_func}
\alias{sc3_assign_func}
\title{Run SC3}
\usage{
sc3_assign_func(sce, out_dir = getwd())
}
\arguments{
\item{sce}{SingleCellExperiment object of the dataset}

\item{out_dir}{Directory where output should be written to. Defaults to the current working directory. Will automatically create a directory called 'alg_clust_assign' to house all algorithm assignments}
}
\description{
Function to run SC3 clustering algorithm.
}
\examples{
sce3_assign_func(sce, out_dir = getwd())
}
\keyword{RaceID}
