% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_clustering_functions.r
\name{scSHC_assign_func}
\alias{scSHC_assign_func}
\title{Run scSHC}
\usage{
scSHC_assign_func(sce, out_dir = getwd(), n_cores = 1)
}
\arguments{
\item{sce}{SingleCellExperiment object of the dataset}

\item{out_dir}{Directory where output should be written to. Defaults to the current working directory. Will automatically create a directory called 'alg_clust_assign' to house all algorithm assignments}

\item{n_cores}{CPU cores for parallel library allocated for running functions.}
}
\description{
Function to run scSHC clustering algorithm.
}
\examples{
scSHC_assign_func(sce, out_dir = getwd())
}
\keyword{scSHC}
