% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_clustering_functions.r
\name{seurat_assign_func}
\alias{seurat_assign_func}
\title{Run Seurat}
\usage{
seurat_assign_func(sce)
}
\arguments{
\item{sce}{SingleCellExperiment object of the dataset}
}
\description{
Function to run Seurat clustering algorithm. Runs Seurat SLC and Seurat Louvain clustering
}
\examples{
seurat_assign_func(sce)
}
\keyword{Seurat}
