% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_clustering_functions.r
\name{spectrum_assign_func}
\alias{spectrum_assign_func}
\title{Run Spectrum}
\usage{
spectrum_assign_func(sce)
}
\arguments{
\item{sce}{SingleCellExperiment object of the dataset}
}
\description{
Function to run Spectrum clustering algorithm
}
\examples{
spectrum_assign_func(sce)
}
\keyword{Spectrum}
